pyproject.toml
[project]
name = "nutriwealth-backend"
version = "1.0.0"
description = "NutriWealth Food & Wellness Tracker Backend API"
authors = [
    {name = "NutriWealth Team", email = "dev@nutriwealth.com"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "requests>=2.31.0",
    "boto3>=1.35.0",
    "openai>=1.54.3",
    "python-dotenv>=1.0.0",
    "pytz>=2024.2",
    "httpx>=0.27.2",
    "PyJWT>=2.8.0",
    "cryptography>=41.0.0",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.5.0",
]

test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "faker>=20.0.0",
    "freezegun>=1.2.0",
]

monitoring = [
    "aws-xray-sdk>=2.12.0",
    "watchtower>=3.0.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["lib", "handlers", "config", "utils", "schemas", "prompts"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "W", "F", "I", "B", "C4", "UP", "SIM"]
ignore = ["E501", "B008", "C901"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = ["-ra", "--strict-markers", "--cov=src"]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
